#82-editTextSetMaxLine BoardAddActivity로 위치
